requirements.txt
streamlit
opencv-python-headless
numpy
insightface
onnxruntime
Pillow